Refactor packaging
